requirements.txt
# === ANTLR runtime for Python ===
antlr4-python3-runtime==4.13.2

# === JSON handling and parsing ===
orjson>=3.10.18 

# === Progress bar (for data processing visualization) ===
tqdm>=4.66

# === File and system utils ===
pathlib

# === TensorFlow ===
tensorflow>=2.13

# === Tools AI model ===
numpy>=1.24
scikit-learn>=1.3.0

# === Tree Visualization ===
anytree>=2.8.0